requirements/base.txt
django==5.2.7
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
drf-spectacular==0.28.0
pendulum==3.1.0
requests==2.32.5
python-decouple==3.8
psycopg2-binary==2.9.11
dj-database-url==3.0.1